Projeto 6 - Deploy de API Para Geração de Texto a Partir de Imagens com LLM/53-Cap08/requirements.txt
fastapi==0.110.0
Pillow==10.0.1
python-multipart==0.0.6
torch==2.2.0
transformers==4.39.1
uvicorn==0.29.0
requests==2.31.0